finialized negotiation functions and endpoints plus re organized structure.
