+++ app/__init__.py
@@ -3,7 +3,6 @@ from flask_sqlalchemy import SQLAlchemy
 from config import Config
 from flasgger import Swagger
 from flask_cors import CORS
-from 
 
 db = SQLAlchemy()
 swagger = Swagger()
@@ -15,20 +14,12 @@ def create_app(config_class=Config):
     db.init_app(app)
     swagger.init_app(app)
 
-    # Registering Models
-    from app.customer.models import Customer
-    from app.inventory.models import Vehical, RetailItem
-    from app.negotiation.models import Negotiation, Offer
-
     # Registering Blueprints
     from app.customer import customer_bp
     app.register_blueprint(customer_bp)
 
     from app.user import user_bp
     app.register_blueprint(user_bp)
-
-    from app.services import services_bp
-    app.register_blueprint(services_bp)
 
     from app.scheduling import scheduling_bp
     app.register_blueprint(scheduling_bp)
@@ -45,27 +36,16 @@ def create_app(config_class=Config):
     from app.inventory import inventory_bp
     app.register_blueprint(inventory_bp)
 
-    from app.routes import routes_bp
-    app.register_blueprint(routes_bp)
-
     # Registering Services
-    from app.services.scheduling_service import ScheduleService
-    app.schedule_service = ScheduleService()
-
-    from app.services.negotiation_service import NegotiationService
+    from app.negotiation.services import NegotiationService
     app.negotiation_service = NegotiationService()
 
-    from app.services.purchasing_service import PurchasingService
-    app.purchasing_service = PurchasingService()
-
-    from app.services.inventory_service import InventoryService
+    from app.inventory.services import InventoryService
     app.inventory_service = InventoryService()
 
     @app.before_request
     def before_request():
-        g.schedule_service = current_app.schedule_service
         g.negotiation_service = current_app.negotiation_service
-        g.purchasing_service = current_app.purchasing_service
         g.inventory_service = current_app.inventory_service
 
     return app
