+++ app/customer/__init__.py
@@ -2,4 +2,4 @@ from flask import Blueprint
 
 customer_bp = Blueprint('customer', __name__, template_folder='templates', static_folder='static')
 
-from . import routes  # Import routes to ensure they're registered with the blueprint
+from . import routes
