+++ app/inventory/__init__.py
@@ -2,3 +2,5 @@ from flask import Blueprint
 
 inventory_bp = Blueprint('inventory', __name__, template_folder='templates', static_folder='static')
 
+from . import routes
+
