+++ app/negotiation/__init__.py
@@ -1,5 +1,3 @@
 from flask import Blueprint
 
 negotiation_bp = Blueprint('negotiation', __name__, template_folder='templates', static_folder='static')
-
-from . import routes  # Import routes to ensure they're registered with the blueprint
