--- app/services/__init__.py
@@ -1,3 +1,0 @@
-from flask import Blueprint
-
-services_bp = Blueprint('services', __name__, template_folder='templates', static_folder='static')
