--- app/services/purchasing_service.py
@@ -1,3 +1,0 @@
-class PurchasingService:
-    # logic for purchasing
-    pass
