--- app/services/scheduling_service.py
@@ -1,3 +1,0 @@
-class ScheduleService:
-    # logic to scheduling
-    pass
